change dtype

+++ source/extensions/omni.isaac.lab/omni/isaac/lab/managers/manager_term_cfg.py
@@ -209,7 +209,7 @@ class EventTermCfg(ManagerTermBaseCfg):
         This is only used if the mode is ``"interval"``.
     """
 
-    min_step_count_between_reset: float = 0.0
+    min_step_count_between_reset: int = 0
     """The minimum number of environment steps between when term is applied.
 
     When mode is "reset", the term will not be applied on the next reset unless
